pythonProject/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.5.30
Django==3.2.7
djangorestframework==3.12.4
mysqlclient==2.0.3
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
wincertstore==0.2
